app/requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
mysql-connector-python==8.1.0
gunicorn==21.2.0
